+++ Source/ASConfigurationInternal.m
@@ -7,6 +7,7 @@
 //
 
 #import "ASConfigurationInternal.h"
+#import <AsyncDisplayKit/ASAssert.h>
 #import <AsyncDisplayKit/ASConfiguration.h>
 #import <AsyncDisplayKit/ASConfigurationDelegate.h>
 #import <stdatomic.h>
@@ -16,6 +17,7 @@
 @implementation ASConfigurationManager {
   ASConfiguration *_config;
   dispatch_queue_t _delegateQueue;
+  BOOL _frameworkInitialized;
   _Atomic(ASExperimentalFeatures) _activatedExperiments;
 }
 
@@ -49,6 +51,21 @@
     }
   }
   return self;
+}
+
+- (void)frameworkDidInitialize
+{
+  ASDisplayNodeAssertMainThread();
+  if (_frameworkInitialized) {
+    ASDisplayNodeFailAssert(@"Framework initialized twice.");
+    return;
+  }
+  _frameworkInitialized = YES;
+  
+  let delegate = _config.delegate;
+  if ([delegate respondsToSelector:@selector(textureDidInitialize)]) {
+    [delegate textureDidInitialize];
+  }
 }
 
 - (BOOL)activateExperimentalFeature:(ASExperimentalFeatures)requested
@@ -89,4 +106,9 @@
 BOOL ASActivateExperimentalFeature(ASExperimentalFeatures feature)
 {
   return [ASGetSharedConfigMgr() activateExperimentalFeature:feature];
+}
+
+void ASNotifyInitialized()
+{
+  [ASGetSharedConfigMgr() frameworkDidInitialize];
 }